% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deptrapr-package.R
\docType{package}
\name{deptrapr-package}
\alias{deptrapr}
\alias{deptrapr-package}
\title{deptrapr: Explore Dependencies Pitfalls}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Test behavior of the different types of depencies with local and CI check.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dagousket/deptrapr}
  \item Report bugs at \url{https://github.com/dagousket/deptrapr/issues}
}

}
\author{
\strong{Maintainer}: Swann Floc'hlay \email{swann@thinkr.fr} (\href{https://orcid.org/0000-0003-1477-830X}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
