% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dataset.R
\name{fetch_dataset}
\alias{fetch_dataset}
\title{fetch_dataset}
\usage{
fetch_dataset(type = c("dino", "penguin"))
}
\arguments{
\item{type}{character Type of dataset to load, can be "dino" or "penguin"}
}
\value{
data.frame A dataset
}
\description{
fetch_dataset
}
\examples{
data <- fetch_dataset(type = "dino")
}
