% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_data}
\alias{show_data}
\title{show_data}
\usage{
show_data(data, view_type = c("head", "skim"))
}
\arguments{
\item{data}{data.frame The dataset to show}

\item{view_type}{character Type of view, show the first rows with "head" or global summary with "skim"}
}
\value{
knitre_kable View the dataset
}
\description{
show_data
}
\examples{
data <- fetch_dataset(type = "dino")
show_data(data = data, view = "skim")
}
