% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{save_plot}
\usage{
save_plot(
  plot,
  ext = c("png", "svg", "jpeg", "pdf"),
  path = "graph_output",
  filename = "output"
)
}
\arguments{
\item{plot}{ggplot A ggplot object to be saved}

\item{ext}{character A vector of output format, can be multiple of "png", "svg", "jpeg", "pdf"}

\item{path}{character A path where to save the output, will be created if it does not exist}

\item{filename}{character The filename fo rthe output, extension will be added}
}
\value{
None Create output files
}
\description{
save_plot
}
\examples{
\dontrun{
save_plot(plot = p,
          filename = "dino")
}
}
